Populate BookFile with a list of section identifers.  This allows a source documentation file (i.e. an xdoc file) to contain multiple sections, and be referenced by the Book Descriptor.
This change modifies the behavior of DefaultBookIo slightly, in that DefaultBookIo now parses documentation files in an attempt to determine the sections they contain.  Furthermore, DefaultBookIo is more strict when parsing files: they have to align with the SiteModule that parses them (e.g. the SiteModule will only parse the files if they sit in the proper directory and carry the proper file extension).

There is also a change of behavior when DefaultBookIndexer (and eventually any other class) when looking up a BookFile for a section.  Section identifiers set on the BookFile are considered before file names as section identifiers.